move min/max to num
